Fall back on get_running_loop in WebSocketCommonProtocol
... to make it compatible with code that does not explicitly provide a loop.
This will be the only pattern supported in the future since in Python >= 3.10 where the
loop argument is removed in the standard lib.

This change was prompted by uvicorn triggering the assertion error because it sets
`self.loop` directly instead of passing it to `super().__init__`.

Loop arg is changed to a kwarg to emphasize that it should not be used.

Fix #988